Extract csv row logic to sanitizeRow
