+++ src/utils/convertCSVToJSON.ts
@@ -1,7 +1,7 @@
 import * as csv from "csv-parser"
 import * as fs from "fs"
-import slugify from "slugify"
 import { Collection } from "../Entities/Collection"
+import { sanitizeRow } from "./sanitizeRow"
 
 export const convertCSVToJSON: (string) => Promise<Collection[]> = (
   path: string
@@ -34,27 +34,21 @@ export const convertCSVToJSON: (string) => Promise<Collection[]> = (
               show_on_editorial = false,
               is_featured_artist_content = false,
             }) =>
-              ({
+              sanitizeRow({
                 title,
-                slug: sanitizeSlug(slug),
+                slug,
                 category,
                 description,
                 headerImage,
                 credit,
-                price_guidance: price_guidance ? Number(price_guidance) : null,
-                show_on_editorial: Boolean(show_on_editorial),
-                is_featured_artist_content: Boolean(is_featured_artist_content),
-                query: (artist_ids || gene_ids || tag_id || keyword) && {
-                  artist_ids: artist_ids
-                    ? artist_ids.split(",").map(a => a.trim())
-                    : [],
-                  gene_ids: gene_ids
-                    ? gene_ids.split(",").map(a => a.trim())
-                    : [],
-                  tag_id,
-                  keyword,
-                },
-              } as Collection)
+                artist_ids,
+                gene_ids,
+                tag_id,
+                keyword,
+                price_guidance,
+                show_on_editorial,
+                is_featured_artist_content,
+              })
           )
 
           resolve(formattedCollections)
@@ -64,13 +58,4 @@ export const convertCSVToJSON: (string) => Promise<Collection[]> = (
       })
       .on("err", reject)
   })
-}
-
-export const sanitizeSlug = (slug: string) => {
-  const cleanedSlug = slugify(slug, {
-    remove: /[.'",&:\/#!$%\^\*;{}=_`’~()]/g,
-    lower: true,
-  })
-
-  return cleanedSlug
 }
